em vez de garantir que não há nulos, colocar let ...

+++ 06_Ex2/app/src/main/java/com/example/a06_ex2/SegundaActivity.kt
@@ -16,13 +16,15 @@ class SegundaActivity : AppCompatActivity() {
         binding = ActivitySegundaBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
-        //alternativa ao bundle, teria é que criar uma variavel j = intent para não perder os valores
+        //alternativa ao bundle, teria é que criar uma variavel n1 = intent para não perder os valores
 
         val i = intent
         val bundle = i.extras
+        //ou: val n1 = i.extras?.getDouble("n1")
+        // e depois em baixo, no buttonOkay: i2.putExtra("n1",n1) ou i2.putExtra(i)
 
         binding.buttonOkay.setOnClickListener {
-            val number2 = binding.editNum.text.toString().toInt()
+            val number2 = binding.editNum.text.toString().toDouble()
             val i2 = Intent(this, TerceiraActivity::class.java)
 
             if (bundle != null) {
